tech(app architecture): Refactor ES5/jQuery code to ES2015/React/Redux
Previously the guts of this were all in one JS file with one accompanying html template that was getting wiley already. The game was not fully implemented when the refactor started; just a few things worked, like generating a random hand from a given deck, clicking on cards to play them on the appropriate row and adding up power of all cards played on a row. The original code is in two files, a JS file and a HTML file, named `initialImplementation` for reference and for extracting some of the logic later as I start reimplementing the interactions that i had working and those for which i had just laid the groundwork.

These changes give the app a more structured architecture and utilizes Grunt/Babel/Webpack to transpile the es2015/esNext code as well as parse React and create a bundle that i can deploy with webpack-dev-server while i develop. Later i will stand it up with an express server, but for now this works. Each of the main pieces of the template were pulled out to create the React components and are mostly in the place where i think they will stay, but lots will probably shift as complexity grows.

Whats working now:
 - Two cards are being manually added to the Redux store
 - CardRowComponent is being populated with two CardComponents and rendering to the page (albeit in a row of zero height since it's being rendered outside of the playfield)

 What's next:
 - put the CardRowComponent in the PlayerFieldComponent then move out from there
 - figure out better way of seeding initial card data in store