feat(player field components): Clean up structure/data flow for player fields
After the last commit, the cards were populating a single row.

Now each of the four types of card rows are populating with the dummy cards as well as the rows for the opposing player, which need to be rendered in reverse order. Now we have data flowing from the outermost container all the way down to the cards within the rows. Later this will be refactored to utilize the `React-Redux` `connect()` method to create container components that have direct access to the data store, instead of having to stitch everything together, like it is now.

Next up:
-reducer(s) for constructing a deck

After that:
-reducers for constructing a hand and storing cards in specific rows.