feat(deck reducer): Create reducer for constructing base decks from card data
Previously cards were being added one at a time directly to the `cardsReducer`. Also, there were several small issues with the logic, where some data was not being assigned to cards or was relying on state for things it did not need to be.

Now we have a `addDecksReducer` that takes in the whole card data object from a static file and compiles the decks. For now it is just pulling in each faction, but it will eventually need to use that data and concat it with the generic cards to fill out the rest of the deck.

Next up:
-reducers for constructing a hand and storing cards in specific rows.

After that:
- (sometime) rename some things and maybe move some reducers into separate files to separate concerns.
- populate hand row with constructed hand