tech(data structures): changed data associations and added counter
The data structure was becoming very hierarchical and data was being duplicated, so I changed nesting of data and data duplication to be based on references to ids of the data in question. players have a reference to a deck id and decks have references to card ids, etc. Card data, also, is no longer duplicated to add state attributes; there are reference, `source` cards and then metadata is associated to each `source` card by id. Currently this metadata is merged with the source card data in a React component, but this could/should occur in a reducer.

The only new functionality this commit is the power counter on the left side of every playfield row. To do this, I moved the container component up to the row level from the card component level so that i would have access not only to the power level of all of the cards for this particular feature, but also so i can more easily handle row effects, like commander's horn or weather effects.

Next up:
- integrate `Immutable.js` and restructure/refactor some things yet again.